+++ src/evm/contract/mod.rs
@@ -1,2 +1,2 @@
-pub mod darkwebb;
+pub mod protocol_solidity;
 pub mod tornado;
